locale.c: Generalize stdsize_locale()
This function is rewritten to handle LC_ALL, and to handle certain buggy
Win32 locale names.  This commit also calls it in appropriate places
where those buggy names could be returned.

setlocale() on Windows may return a locale that cannot be used as input
to a future setlocale().  This is contrary to the C89 standard, and
appears to have been an oversight corrected in the most recent Windows
version(s).

This commit solves the problem (as far as I know) by looking for the
problematic syntax and adjusting it.

I also rewrote the function to handle LC_ALL, which fixes that deficiency.

And, a change in that that I think is an improvement is that everything
starting with a \n is trimmed, instead of just a trailing \n being
chomped.